json merged from master, with the docker fix
