+++ docker/base/Dockerfile
@@ -1,1 +1,3 @@
-FROM redis:6.2.5-buster
+FROM redis:6.2.6-buster
+
+CMD ["redis-server", "/redis.conf"]
