+++ docker/base/Dockerfile.sentinel
@@ -1,0 +1,3 @@
+FROM redis:6.2.6-buster
+
+CMD ["redis-sentinel", "/sentinel.conf"]
