+++ docker/base/README.md
@@ -1,0 +1,1 @@
+Dockers in this folder are built, and uploaded to the redisfab dockerhub store.
