+++ docker/master/redis.conf
@@ -1,3 +1,2 @@
-bind master 127.0.0.1
 port 6379
 save ""
