--- docker/replica/Dockerfile
@@ -1,7 +1,0 @@
-FROM redis-py-base:latest
-
-COPY redis.conf /usr/local/etc/redis/redis.conf
-
-EXPOSE 6380
-
-CMD [ "redis-server", "/usr/local/etc/redis/redis.conf" ]
