+++ docker/replica/redis.conf
@@ -1,4 +1,3 @@
-bind replica 127.0.0.1
 port 6380
 save ""
 replicaof master 6379
