--- docker/sentinel_1/Dockerfile
@@ -1,7 +1,0 @@
-FROM redis-py-base:latest
-
-COPY sentinel.conf /usr/local/etc/redis/sentinel.conf
-
-EXPOSE 26379
-
-CMD [ "redis-sentinel", "/usr/local/etc/redis/sentinel.conf" ]
