+++ docker/sentinel_2/sentinel.conf
@@ -1,7 +1,6 @@
-bind sentinel_2 127.0.0.1
 port 26380
 
-sentinel monitor redis-py-test master 6379 2
+sentinel monitor redis-py-test 127.0.0.1 6379 2
 sentinel down-after-milliseconds redis-py-test 5000
 sentinel failover-timeout redis-py-test 60000
 sentinel parallel-syncs redis-py-test 1
