+++ tox.ini
@@ -10,56 +10,56 @@ envlist = {py35,py36,py37,py38,py39,pypy3}-{plain,hiredis}, flake8
 
 [docker:master]
 name = master
-image = redis:6.2-bullseye
+image = redisfab/redis-py:6.2.6-buster
 ports =
     6379:6379/tcp
 healtcheck_cmd = python -c "import socket;print(True) if 0 == socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect_ex(('127.0.0.1',6379)) else False"
 volumes =
-    bind:rw:{toxinidir}/docker/master/redis.conf:/usr/local/etc/redis/redis.conf
+    bind:rw:{toxinidir}/docker/master/redis.conf:/redis.conf
 
 [docker:replica]
 name = replica
-image = redis:6.2-bullseye
+image = redisfab/redis-py:6.2.6-buster
 links =
     master:master
 ports =
     6380:6380/tcp
 healtcheck_cmd = python -c "import socket;print(True) if 0 == socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect_ex(('127.0.0.1',6380)) else False"
 volumes =
-    bind:rw:{toxinidir}/docker/replica/redis.conf:/usr/local/etc/redis/redis.conf
+    bind:rw:{toxinidir}/docker/replica/redis.conf:/redis.conf
 
 [docker:sentinel_1]
 name = sentinel_1
-image = redis:6.2-bullseye
+image = redisfab/redis-py-sentinel:6.2.6-buster
 links =
     master:master
 ports =
     26379:26379/tcp
 healtcheck_cmd = python -c "import socket;print(True) if 0 == socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect_ex(('127.0.0.1',26379)) else False"
 volumes =
-    bind:rw:{toxinidir}/docker/sentinel_1/sentinel.conf:/usr/local/etc/redis/sentinel.conf
+    bind:rw:{toxinidir}/docker/sentinel_1/sentinel.conf:/sentinel.conf
 
 [docker:sentinel_2]
 name = sentinel_2
-image = redis:6.2-bullseye
+image = redisfab/redis-py-sentinel:6.2.6-buster
 links =
     master:master
 ports =
     26380:26380/tcp
 healtcheck_cmd = python -c "import socket;print(True) if 0 == socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect_ex(('127.0.0.1',26380)) else False"
 volumes =
-    bind:rw:{toxinidir}/docker/sentinel_2/sentinel.conf:/usr/local/etc/redis/sentinel.conf
+    bind:rw:{toxinidir}/docker/sentinel_2/sentinel.conf:/sentinel.conf
 
 [docker:sentinel_3]
 name = sentinel_3
-image = redis:6.2-bullseye
+image = redisfab/redis-py-sentinel:6.2.6-buster
 links =
     master:master
 ports =
     26381:26381/tcp
 healtcheck_cmd = python -c "import socket;print(True) if 0 == socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect_ex(('127.0.0.1',26381)) else False"
 volumes =
-    bind:rw:{toxinidir}/docker/sentinel_3/sentinel.conf:/usr/local/etc/redis/sentinel.conf
+    bind:rw:{toxinidir}/docker/sentinel_3/sentinel.conf:/sentinel.conf
 
 [docker:redismod]
 name = redismod
